split classes into more files
