+++ .gitignore
@@ -37,9 +37,12 @@ Thumbs.db
 */bin/
 */obj/
 
-/Spoke.Runtime.cs.meta
-/Spoke.Reactive.cs.meta
-/Spoke.Unity.cs.meta
+/Spoke.Runtime.meta
+/Spoke.Runtime/**/*.meta
+/Spoke.Reactive.meta
+/Spoke.Reactive/**/*.meta
+/Spoke.Unity.meta
+/Spoke.Unity/**/*.meta
 /LICENSE.meta
 /README.md.meta
 /CHANGELOG.md.meta
